# Parse the header to get meta data about the language