--- a/agent/agent.go
+++ b/agent/agent.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"sync"
-	"sync/atomic"
 	"time"
 
 	log "github.com/cihub/seelog"
@@ -139,8 +138,8 @@ func (a *Agent) Process(t model.Trace) {
 	root := t.GetRoot()
 	if root.End() < model.Now()-2*a.conf.BucketInterval.Nanoseconds() {
 		log.Errorf("skipping trace with root too far in past, root:%v", *root)
-		atomic.AddInt64(&a.Receiver.stats.TracesDropped, 1)
-		atomic.AddInt64(&a.Receiver.stats.SpansDropped, int64(len(t)))
+		//atomic.AddInt64(&a.Receiver.stats.TracesDropped, 1)
+		//atomic.AddInt64(&a.Receiver.stats.SpansDropped, int64(len(t)))
 		return
 	}