--- a/agent/agent.go
+++ b/agent/agent.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"sync"
+	"sync/atomic"
 	"time"
 
 	log "github.com/cihub/seelog"
@@ -139,10 +140,9 @@ func (a *Agent) Process(t model.Trace) {
 	if root.End() < model.Now()-2*a.conf.BucketInterval.Nanoseconds() {
 		log.Errorf("skipping trace with root too far in past, root:%v", *root)
 		tags := []string{"method:Process"}
-		ts := newTagStats(tags)
-		ts.TracesDropped++
-		ts.SpansDropped += int64(len(t))
-		a.Receiver.stats.update(ts)
+		ts := a.Receiver.stats.getTagStats(tags)
+		atomic.AddInt64(&ts.TracesDropped, 1)
+		atomic.AddInt64(&ts.SpansDropped, int64(len(t)))
 		return
 	}