--- a/agent/receiver.go
+++ b/agent/receiver.go
@@ -342,6 +342,14 @@ func decodeReceiverPayload(r io.Reader, dest msgp.Decodable, v APIVersion, conte
 	}
 }
 
+// headerFields is a map used to decode the header metas
+var headerFields = map[string]string{
+	"lang":           "Datadog-Meta-Lang",
+	"lang_version":   "Datadog-Meta-Lang-Version",
+	"interpreter":    "Datadog-Meta-Lang-Interpreter",
+	"tracer_version": "Datadog-Meta-Tracer-Version",
+}
+
 // parseTags extracts tags from the header request
 func parseTags(req *http.Request) []string {
 	tags := []string{}