… agent/agent_test.go
… agent/info_test.go
… agent/receiver_test.go
… agent/stats.go