- [x] Deployed on staging

New tags available for trace-agent metrics:
- `lang`
- `lang_version`
- `interpreter`
- `trace_version`
- `endpoint`

## Benchmarks

- [master version](https://github.com/DataDog/datadog-trace-agent/commit/1a2473c47d170054325c9998836939a4ae0c23fb): `go test -run=NONE -bench=. > master.txt`

```
BenchmarkAgentTraceProcessing-4           	   10000	    157463 ns/op	   50315 B/op	     705 allocs/op
BenchmarkWatchdog-4                       	 3000000	       477 ns/op	      48 B/op	       1 allocs/op
BenchmarkEncodeAgentPayload-4             	     200	   9962172 ns/op	 2183083 B/op	   21829 allocs/op
BenchmarkHandleSpanRandom-4               	   20000	     65146 ns/op	   28907 B/op	     381 allocs/op
BenchmarkHandleTracesFromOneApp-4         	  200000	      7630 ns/op	    1362 B/op	      22 allocs/op
BenchmarkHandleTracesFromMultipleApps-4   	  200000	      8303 ns/op	    1362 B/op	      22 allocs/op
BenchmarkDecoderJSON-4                    	      10	 121390220 ns/op	25726412 B/op	  329443 allocs/op
BenchmarkDecoderMsgpack-4                 	     100	  20649946 ns/op	 9225670 B/op	  118953 allocs/op
```

- new version: `go test -run=NONE -bench=. > meta.txt`

```
BenchmarkAgentTraceProcessing-4           	   10000	    142788 ns/op	   50233 B/op	     705 allocs/op
BenchmarkWatchdog-4                       	 3000000	       459 ns/op	      48 B/op	       1 allocs/op
BenchmarkEncodeAgentPayload-4             	     200	   6760406 ns/op	 2188907 B/op	   21830 allocs/op
BenchmarkHandleSpanRandom-4               	   20000	     65688 ns/op	   28907 B/op	     381 allocs/op
BenchmarkHandleTracesFromOneApp-4         	  200000	      8342 ns/op	    1362 B/op	      22 allocs/op
BenchmarkHandleTracesFromMultipleApps-4   	  200000	      8682 ns/op	    1362 B/op	      22 allocs/op
BenchmarkDecoderJSON-4                    	      10	 117905562 ns/op	25726409 B/op	  329443 allocs/op
BenchmarkDecoderMsgpack-4                 	     100	  21261181 ns/op	 9225620 B/op	  118953 allocs/op
```

- diff: `benchcmp master.txt meta.txt`

```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkAgentTraceProcessing-4             157463        142788        -9.32%
BenchmarkWatchdog-4                         477           459           -3.77%
BenchmarkEncodeAgentPayload-4               9962172       6760406       -32.14%
BenchmarkHandleSpanRandom-4                 65146         65688         +0.83%
BenchmarkHandleTracesFromOneApp-4           7630          8342          +9.33%
BenchmarkHandleTracesFromMultipleApps-4     8303          8682          +4.56%
BenchmarkDecoderJSON-4                      121390220     117905562     -2.87%
BenchmarkDecoderMsgpack-4                   20649946      21261181      +2.96%

benchmark                                   old allocs     new allocs     delta
BenchmarkAgentTraceProcessing-4             705            705            +0.00%
BenchmarkWatchdog-4                         1              1              +0.00%
BenchmarkEncodeAgentPayload-4               21829          21830          +0.00%
BenchmarkHandleSpanRandom-4                 381            381            +0.00%
BenchmarkHandleTracesFromOneApp-4           22             22             +0.00%
BenchmarkHandleTracesFromMultipleApps-4     22             22             +0.00%
BenchmarkDecoderJSON-4                      329443         329443         +0.00%
BenchmarkDecoderMsgpack-4                   118953         118953         +0.00%

benchmark                                   old bytes     new bytes     delta
BenchmarkAgentTraceProcessing-4             50315         50233         -0.16%
BenchmarkWatchdog-4                         48            48            +0.00%
BenchmarkEncodeAgentPayload-4               2183083       2188907       +0.27%
BenchmarkHandleSpanRandom-4                 28907         28907         +0.00%
BenchmarkHandleTracesFromOneApp-4           1362          1362          +0.00%
BenchmarkHandleTracesFromMultipleApps-4     1362          1362          +0.00%
BenchmarkDecoderJSON-4                      25726412      25726409      -0.00%
BenchmarkDecoderMsgpack-4                   9225670       9225620       -0.00%
```