@gabsn 2017-07-05
For the moment the tests break (because I didn't modify the info part of the agent). @palazzem can you check if the global implementation satisfies you ?